Cut-paste connection related code to separate file
Preparation for making this a module
